+++ Automate/Framework/Storage/ContainerExtensions.cs
@@ -12,28 +12,28 @@ namespace Pathoschild.Stardew.Automate.Framework.Storage
         /// <param name="container">The container instance.</param>
         public static bool StorageAllowed(this IContainer container)
         {
-            return !container.ShouldIgnore() && !container.HasTag("automate:nooutput");
+            return !container.HasTag("automate:no-store");
         }
 
         /// <summary>Get whether this container should be preferred when choosing where to store items.</summary>
         /// <param name="container">The container instance.</param>
         public static bool StoragePreferred(this IContainer container)
         {
-            return container.StorageAllowed() && container.HasTag("automate:output");
+            return container.StorageAllowed() && container.HasTag("automate:prefer-store");
         }
 
         /// <summary>Get whether items can be retrieved from this container.</summary>
         /// <param name="container">The container instance.</param>
         public static bool TakingItemsAllowed(this IContainer container)
         {
-            return !container.ShouldIgnore() && !container.HasTag("automate:noinput");
+            return !container.HasTag("automate:no-take");
         }
 
         /// <summary>Get whether this container should be preferred when choosing where to retrieve items.</summary>
         /// <param name="container">The container instance.</param>
         public static bool TakingItemsPreferred(this IContainer container)
         {
-            return container.TakingItemsAllowed() && container.HasTag("automate:input");
+            return container.TakingItemsAllowed() && container.HasTag("automate:prefer-take");
         }
 
 
@@ -46,14 +46,6 @@ namespace Pathoschild.Stardew.Automate.Framework.Storage
         private static bool HasTag(this IContainer container, string tag)
         {
             return container.Name?.IndexOf($"|{tag}|", StringComparison.InvariantCultureIgnoreCase) >= 0;
-        }
-
-        /// <summary>Get whether this container should be preferred for output when possible.</summary>
-        /// <param name="container">The container instance.</param>
-        private static bool ShouldIgnore(this IContainer container)
-        {
-            // legacy tag no longer set as of Chests Anywhere 1.17.3, but some players may still have it in their save
-            return container.HasTag("automate:ignore");
         }
     }
 }