+++ Common/Integrations/ProducerFrameworkMod/ProducerFrameworkModIntegration.cs
@@ -1,4 +1,6 @@
+using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using StardewModdingAPI;
 using SObject = StardewValley.Object;
@@ -13,6 +15,9 @@ namespace Pathoschild.Stardew.Common.Integrations.ProducerFrameworkMod
         *********/
         /// <summary>The mod's public API.</summary>
         private readonly IProducerFrameworkModApi ModApi;
+
+        /// <summary>Whether the integration has already logged a warning caused by an invalid recipe.</summary>
+        private bool LoggedInvalidRecipeError = false;
 
 
         /*********
@@ -38,7 +43,7 @@ namespace Pathoschild.Stardew.Common.Integrations.ProducerFrameworkMod
         {
             this.AssertLoaded();
 
-            return this.ModApi.GetRecipes().Select(this.ReadRecipe);
+            return this.ReadRecipes(this.ModApi.GetRecipes());
         }
 
         /// <summary>Get the list of recipes for a machine.</summary>
@@ -48,28 +53,65 @@ namespace Pathoschild.Stardew.Common.Integrations.ProducerFrameworkMod
         {
             this.AssertLoaded();
 
-            return this.ModApi.GetRecipes(machine).Select(this.ReadRecipe);
+            return this.ReadRecipes(this.ModApi.GetRecipes(machine));
         }
 
 
         /*********
         ** Private methods
         *********/
+        /// <summary>Read the metadata for recipes provided by <see cref="GetRecipes()"/> or <see cref="GetRecipes(SObject)"/>.</summary>
+        /// <param name="raw">The raw recipe data.</param>
+        private IEnumerable<ProducerFrameworkRecipe> ReadRecipes(IEnumerable<IDictionary<string, object>> raw)
+        {
+            return raw
+                .Select(this.ReadRecipe)
+                .Where(p => p != null);
+        }
+
+
         /// <summary>Read the metadata for a recipe provided by <see cref="GetRecipes()"/> or <see cref="GetRecipes(SObject)"/>.</summary>
         /// <param name="raw">The raw recipe data.</param>
+        [SuppressMessage("ReSharper", "UseObjectOrCollectionInitializer", Justification = "Avoid object initializer so it's clear which line failed in stack traces reported by players, since we don't control the format being parsed.")]
         private ProducerFrameworkRecipe ReadRecipe(IDictionary<string, object> raw)
         {
-            ProducerFrameworkRecipe recipe = new ProducerFrameworkRecipe();
-            recipe.InputId = raw["InputKey"] as int?;
-            recipe.MachineId = (int)raw["MachineID"];
-            recipe.Ingredients = ((List<Dictionary<string, object>>)raw["Ingredients"]).ToDictionary(p => (int)p["ID"], p => (int)p["Count"]);
-            recipe.ExceptIngredients = ((List<Dictionary<string, object>>)raw["ExceptIngredients"]).Select(p => (int)p["ID"]).ToArray();
-            recipe.OutputId = (int)raw["Output"];
-            recipe.MinOutput = (int)raw["MinOutput"];
-            recipe.MaxOutput = (int)raw["MaxOutput"];
-            recipe.PreserveType = (SObject.PreserveType?)raw["PreserveType"];
-            recipe.OutputChance = (double)raw["OutputChance"];
-            return recipe;
+            try
+            {
+                ProducerFrameworkRecipe recipe = new ProducerFrameworkRecipe();
+                recipe.InputId = raw["InputKey"] as int?;
+                recipe.MachineId = (int)raw["MachineID"];
+                recipe.Ingredients = ((List<Dictionary<string, object>>)raw["Ingredients"]).Select(this.ReadIngredient).ToArray();
+                recipe.ExceptIngredients = ((List<Dictionary<string, object>>)raw["ExceptIngredients"]).Select(this.ReadIngredient).Select(p => p.InputId).ToArray();
+                recipe.OutputId = (int)raw["Output"];
+                recipe.MinOutput = (int)raw["MinOutput"];
+                recipe.MaxOutput = (int)raw["MaxOutput"];
+                recipe.PreserveType = (SObject.PreserveType?)raw["PreserveType"];
+                recipe.OutputChance = (double)raw["OutputChance"];
+                return recipe;
+            }
+            catch (Exception ex)
+            {
+                if (!this.LoggedInvalidRecipeError)
+                {
+                    this.LoggedInvalidRecipeError = true;
+                    this.Monitor.Log("Failed to load some recipes from Producer Framework Mod. Some custom machines may not appear in lookups.", LogLevel.Warn);
+                    this.Monitor.Log(ex.ToString());
+                }
+                return null;
+            }
+        }
+
+        /// <summary>Read an ingredient if it's supported.</summary>
+        /// <param name="raw">The raw ingredient model.</param>
+        private ProducerFrameworkIngredient ReadIngredient(IDictionary<string, object> raw)
+        {
+            int? id = int.TryParse(raw["ID"]?.ToString(), out int parsedId)
+                ? parsedId
+                : null as int?;
+            int count = raw.TryGetValue("Count", out object rawCount)
+                ? (int)rawCount
+                : 1;
+            return new ProducerFrameworkIngredient { InputId = id, Count = count };
         }
     }
 }